fix(monitor): remove useless code
